autopilot/agent: move choice algorithms to new file choice.go
